Fix type errors of the arguments for format strings [-Wformat=]
- Cast the results of `sizeof` and `strlen()` to `int` for the field width `*`
- Cast `signed` types to `unsigned` types for the conversion specifier `x`
- Use the conversion specifier `u` for printing `unsigned` types as decimal integers
- Use the conversion specifier `d` for printing `signed` types as decimal integers
- Use the length modifier `l` for `long`, `time_t`, `off_t`, and `key_t`
- Use the length modifier `z` for `size_t`

(cherry picked from commit d02ac86ef33725be9a3e33b5b9ead0fdc830ad16)

+++ maple/xchatd.c
@@ -1112,7 +1112,7 @@ chat_list_rooms(
             }
             else
             {
-                sprintf(from, "%%-%ds¢x%%4d¢x%%s", strlen(CHATROOMNAME)-6+12);
+                sprintf(from, "%%-%ds¢x%%4d¢x%%s", (int)(unsigned int)strlen(CHATROOMNAME)-6+12);
                 sprintf(buf, from, cr->name, cr->occupants, cr->topic);
                 if (LOCKED(cr))
                     strcat(buf, " [Âê¦í]");
@@ -3118,8 +3118,8 @@ server_usage(int signum)
     fprintf(flog, "\n[Server Usage]\n\n"
         "user time: %.6f\n"
         "system time: %.6f\n"
-        "maximum resident set size: %lu P\n"
-        "integral resident set size: %lu\n"
+        "maximum resident set size: %ld P\n"
+        "integral resident set size: %ld\n"
         "page faults not requiring physical I/O: %d\n"
         "page faults requiring physical I/O: %d\n"
         "swaps: %d\n"
